chore: folder restructure + fix corelayout setup + authcontext
