+++ apps/frontend/app/(auth)/auth/register/page.tsx
@@ -1,18 +1,17 @@
-import { type Metadata } from "next";
 import Link from "next/link";
 
-import { cn } from "../../../lib/utils";
-import { buttonVariants } from "../../../components/ui/button";
-import { Icons } from "../../../components/icons";
-import { LoginForm } from "../../../components/login";
+import { cn } from "../../../../lib/utils";
+import { buttonVariants } from "../../../../components/ui/button";
+import { SignupForm } from "../../../../components/auth/signup";
 import Image from "next/image";
+import { Icons } from "@/components/icons";
 
-export const metadata: Metadata = {
-  title: "Login",
-  description: "Login to your account",
+export const metadata = {
+  title: "Create an account",
+  description: "Create an account to get started.",
 };
 
-export default function LoginPage() {
+export default function RegisterPage() {
   return (
     <div className="container flex h-screen w-screen flex-col items-center justify-center">
       <Link
@@ -27,25 +26,26 @@ export default function LoginPage() {
           Home
         </>
       </Link>
+      <Link
+        href="/auth/login"
+        className={cn(
+          buttonVariants({ variant: "default" }),
+          "absolute right-4 top-4 md:right-8 md:top-8"
+        )}
+      >
+        Login
+      </Link>
       <div className="mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]">
         <div className="flex flex-col space-y-2 text-center items-center justify-center">
           <Image src={"../logo.svg"} alt="logo" width={75} height={75} />
           <h1 className="text-2xl font-semibold tracking-tight">
-            Welcome back
+            Create an account
           </h1>
           <p className="text-sm text-muted-foreground">
-            Enter your credentials to sign in to your account
+            Enter your credentials below to create your account
           </p>
         </div>
-        <LoginForm />
-        <p className="px-8 text-center text-sm text-muted-foreground">
-          <Link
-            href="/auth/register"
-            className="hover:text-brand underline underline-offset-4"
-          >
-            Don&apos;t have an account? Sign Up
-          </Link>
-        </p>
+        <SignupForm />
       </div>
     </div>
   );
