+++ apps/frontend/app/(core)/page.tsx
@@ -1,0 +1,10 @@
+import { Button } from "../../components/ui/button";
+import { navConfig } from "../../config/nav";
+import { MainNav } from "../../components/main-nav";
+import { UserAccountNav } from "../../components/user-account-nav";
+import Link from "next/link";
+import { SiteFooter } from "../../components/site-footer";
+
+export default async function Home() {
+  return <></>;
+}
