+++ apps/frontend/app/(dashboard)/dashboard/projects/loading.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { CardSkeleton } from "../../../components/card-skeleton";
-import { DashboardHeader } from "../../../components/header";
-import { DashboardShell } from "../../../components/shell";
-import { Card } from "../../../components/ui/card";
+import { CardSkeleton } from "../../../../components/card-skeleton";
+import { DashboardHeader } from "../../../../components/header";
+import { DashboardShell } from "../../../../components/shell";
+import { Card } from "../../../../components/ui/card";
 
 export default function DashboardProjectsLoading() {
   return (
