+++ apps/frontend/components/auth/auth-provider.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import { createContext, useContext, useEffect, useState } from "react";
+import { User } from "../../types/user.types";
+
+type AuthState = {
+  user?: User;
+};
+export const AuthContext = createContext<AuthState>({});
+
+export function useAuth() {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error("useAuth must be used within an AuthProvider");
+  }
+  return context;
+}
+
+export function AuthProvider({ children }: { children: React.ReactNode }) {
+  const [user, setUser] = useState<User>();
+
+  useEffect(() => {
+    const auth = localStorage.getItem("auth");
+    if (auth) {
+      setUser(JSON.parse(auth));
+    }
+  }, []);
+
+  return (
+    <AuthContext.Provider value={{ user }}>{children}</AuthContext.Provider>
+  );
+}
