+++ apps/frontend/components/auth/login.tsx
@@ -1,0 +1,108 @@
+"use client";
+
+import * as React from "react";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import type * as z from "zod";
+
+import { cn } from "../../lib/utils";
+import { userLoginSchema } from "../../lib/validations/auth";
+import { buttonVariants } from "../ui/button";
+import { Input } from "../ui/input";
+import { Label } from "../ui/label";
+import { Icons } from "../icons";
+import { login } from "@/lib/auth";
+import { toast } from "../ui/use-toast";
+import { useRouter } from "next/navigation";
+
+type LoginFormProps = React.HTMLAttributes<HTMLDivElement>;
+
+type FormData = z.infer<typeof userLoginSchema>;
+
+export function LoginForm({ className, ...props }: LoginFormProps) {
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<FormData>({
+    resolver: zodResolver(userLoginSchema),
+  });
+  const router = useRouter();
+
+  const [isLoading, setIsLoading] = React.useState<boolean>(false);
+
+  async function onSubmit(data: FormData) {
+    setIsLoading(true);
+    const { email, password } = data;
+    const response = await login({ email, password });
+    setIsLoading(false);
+
+    if (response?.status !== 200) {
+      return toast({
+        title: "Something went wrong.",
+        description: response.message,
+        variant: "destructive",
+      });
+    }
+
+    toast({
+      description: "Welcome back!",
+    });
+
+    router.push("/dashboard");
+  }
+
+  return (
+    <div className={cn("grid gap-6", className)} {...props}>
+      <form onSubmit={handleSubmit(onSubmit)}>
+        <div className="grid gap-2">
+          <div className="grid gap-1">
+            <Label className="sr-only" htmlFor="email">
+              Email
+            </Label>
+            <Input
+              id="email"
+              placeholder="name@example.com"
+              type="email"
+              autoCapitalize="none"
+              autoComplete="email"
+              autoCorrect="off"
+              disabled={isLoading}
+              {...register("email")}
+            />
+            {errors?.email && (
+              <p className="px-1 text-xs text-red-600">
+                {errors.email.message}
+              </p>
+            )}
+          </div>
+          <div className="grid gap-1">
+            <Label className="sr-only" htmlFor="password">
+              Password
+            </Label>
+            <Input
+              id="password"
+              placeholder="********"
+              type="password"
+              autoCapitalize="none"
+              autoCorrect="off"
+              disabled={isLoading}
+              {...register("password")}
+            />
+            {errors?.password && (
+              <p className="px-1 text-xs text-red-600">
+                {errors.password.message}
+              </p>
+            )}
+          </div>
+          <button className={cn(buttonVariants())} disabled={isLoading}>
+            {isLoading && (
+              <Icons.spinner className="mr-2 h-4 w-4 animate-spin" />
+            )}
+            Sign In
+          </button>
+        </div>
+      </form>
+    </div>
+  );
+}
